PEPPER-1402 changed BaseKitTestUtil into KitTestUtil and removed PepperKitUtil.java and JuniperKitUtil.java
